Move PostsFilterUtils to core:data module
