+++ shared/src/commonMain/kotlin/dev/sasikanth/rss/reader/feed/FeedPresenter.kt
@@ -24,7 +24,7 @@ import dev.sasikanth.rss.reader.data.repository.ObservableActiveSource
 import dev.sasikanth.rss.reader.data.repository.RssRepository
 import dev.sasikanth.rss.reader.data.repository.SettingsRepository
 import dev.sasikanth.rss.reader.data.time.LastRefreshedAt
-import dev.sasikanth.rss.reader.posts.PostsFilterUtils
+import dev.sasikanth.rss.reader.data.utils.PostsFilterUtils
 import dev.sasikanth.rss.reader.util.DispatchersProvider
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
